bms/requirements.txt
BAC0==23.7.3
pytz==2023.3.post1
mysql-connector-python==8.1.0
